+++ lab_2/src/main/kotlin/it/polito/ecommerce/services/WalletService.kt
@@ -8,11 +8,11 @@ import java.math.BigDecimal
 import java.sql.Timestamp
 
 interface WalletService {
-    fun addWallet(customerID: Int): WalletDTO
-    fun getWallet(walletID: Int): WalletDTO
-    fun performTransaction(senderID: Int, receiverID: Int, amount: BigDecimal): TransactionDTO
-//    fun getWalletTransactions(walletID: Int): List<TransactionDTO>
-    fun getWalletTransactions(walletID: Int,
+    fun addWallet(customerID: Long): WalletDTO
+    fun getWallet(walletID: Long): WalletDTO
+    fun performTransaction(senderID: Long, receiverID: Long, amount: BigDecimal): TransactionDTO
+//    fun getWalletTransactions(walletID: Long): List<TransactionDTO>
+    fun getWalletTransactions(walletID: Long,
                                          from: Long? = null,
                                          to: Long? = null)
     : List<TransactionDTO>
